+++ app/src/main/java/net/exent/flywithme/fragment/TakeoffDetails.java
@@ -117,7 +117,7 @@ public class TakeoffDetails extends Fragment {
             public void onClick(View v) {
                 Intent intent = new Intent(getActivity(), FlyWithMeService.class);
                 intent.setAction(FlyWithMeService.ACTION_GET_METEOGRAM);
-                intent.putExtra(FlyWithMeService.ARG_TAKEOFF_ID, (long) takeoff.getId());
+                intent.putExtra(FlyWithMeService.ARG_TAKEOFF_ID, takeoff.getId());
                 getActivity().startService(intent);
                 // show loading animation
                 ProgressBar progressBar = (ProgressBar) getActivity().findViewById(R.id.progressBar2);
